+++ src/core/src/lib/core.module.ts
@@ -11,6 +11,7 @@ import { FormlyTemplateType } from './templates/field-template.type';
 
 import { FieldExpressionExtension } from './extensions/field-expression/field-expression';
 import { FieldValidationExtension } from './extensions/field-validation/field-validation';
+import { FieldFormExtension } from './extensions/field-form/field-form';
 
 export function defaultFormlyConfig(formlyConfig: FormlyConfig): ConfigOption {
   return {
@@ -20,6 +21,7 @@ export function defaultFormlyConfig(formlyConfig: FormlyConfig): ConfigOption {
     ],
     extensions: [
       { name: 'field-validation', extension: new FieldValidationExtension(formlyConfig) },
+      { name: 'field-form', extension: new FieldFormExtension() },
       { name: 'field-expression', extension: new FieldExpressionExtension() },
     ],
   };
